@@ -9,15 +9,12 @@ import java.util.Map;
 
 import javax.sql.DataSource;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.jdbc.EmbeddedDatabaseConnection;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import com.creactiviti.piper.core.context.JdbcContextRepository;
 import com.creactiviti.piper.core.job.JdbcJobRepository;
@@ -38,8 +35,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.google.common.collect.ImmutableMap;
 
-@RunWith(SpringRunner.class)
-@AutoConfigureTestDatabase(connection = EmbeddedDatabaseConnection.H2)
+
+@SpringBootTest
 public class CoordinatorTests {
 
   @Autowired
@@ -115,7 +112,7 @@ public class CoordinatorTests {
     
     Job completedJob = jobRepository.findOne(job.getId());
     
-    Assert.assertEquals(JobStatus.COMPLETED, completedJob.getStatus());
+    Assertions.assertEquals(JobStatus.COMPLETED, completedJob.getStatus());
   }
 
   private ObjectMapper createObjectMapper() {
@@ -125,11 +122,13 @@ public class CoordinatorTests {
     return objectMapper;
   }
   
-  @Test(expected=IllegalArgumentException.class)
+  @Test
   public void testRequiredParams () {
-    Coordinator coordinator = new Coordinator ();
-    coordinator.setPipelineRepository(new ResourceBasedPipelineRepository());
-    coordinator.create(MapObject.of(Collections.singletonMap("pipelineId","demo/hello")));
+    Assertions.assertThrows(IllegalArgumentException.class, () -> {
+      Coordinator coordinator = new Coordinator ();
+      coordinator.setPipelineRepository(new ResourceBasedPipelineRepository());
+      coordinator.create(MapObject.of(Collections.singletonMap("pipelineId","demo/hello")));
+    });
   }
   
 }
