@@ -1,7 +1,7 @@
 
 package com.creactiviti.piper.core.task;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -10,8 +10,8 @@ import static org.mockito.Mockito.when;
 import java.util.Arrays;
 import java.util.Collections;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.mockito.ArgumentCaptor;
 
 import com.creactiviti.piper.core.context.ContextRepository;
@@ -36,7 +36,7 @@ public class SwitchTaskDispatcherTests {
     dispatcher.dispatch(task);
     ArgumentCaptor<TaskExecution> argument = ArgumentCaptor.forClass(TaskExecution.class);
     verify(taskDispatcher,times(1)).dispatch(argument.capture());
-    Assert.assertEquals("print", argument.getValue().getType());
+    Assertions.assertEquals("print", argument.getValue().getType());
   }
   
   @Test
@@ -63,7 +63,7 @@ public class SwitchTaskDispatcherTests {
     dispatcher.dispatch(task);
     ArgumentCaptor<TaskExecution> argument = ArgumentCaptor.forClass(TaskExecution.class);
     verify(taskDispatcher,times(1)).dispatch(argument.capture());
-    Assert.assertEquals("sleep", argument.getValue().getType());
+    Assertions.assertEquals("sleep", argument.getValue().getType());
   }
   
   @Test
@@ -81,7 +81,7 @@ public class SwitchTaskDispatcherTests {
     ArgumentCaptor<String> arg1 = ArgumentCaptor.forClass(String.class);
     ArgumentCaptor<TaskExecution> arg2 = ArgumentCaptor.forClass(TaskExecution.class);
     verify(messenger,times(1)).send(arg1.capture(),arg2.capture());
-    Assert.assertEquals("1234", arg2.getValue().getOutput());
+    Assertions.assertEquals("1234", arg2.getValue().getOutput());
   }
   
 }
