@@ -1,8 +1,9 @@
 
 package com.creactiviti.piper.core.task;
 
-import org.junit.Assert;
-import org.junit.Test;
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import com.creactiviti.piper.core.messenger.Queues;
 
@@ -10,7 +11,7 @@ public class WorkTaskExecutorTests {
 
   @Test
   public void test1 () {
-    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assert.assertEquals(Queues.TASKS, k));
+    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assertions.assertEquals(Queues.TASKS, k));
     executor.dispatch(SimpleTaskExecution.create());
   }
   
@@ -18,7 +19,7 @@ public class WorkTaskExecutorTests {
   public void test2 () {
     SimpleTaskExecution task = SimpleTaskExecution.create();
     task.setNode("encoder");
-    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assert.assertEquals("encoder", k));
+    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assertions.assertEquals("encoder", k));
     executor.dispatch(task);
   }
   
@@ -26,7 +27,7 @@ public class WorkTaskExecutorTests {
   public void test3 () {
     SimpleTaskExecution task = SimpleTaskExecution.create();
     task.setNode("encoder.xlarge");
-    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assert.assertEquals("encoder.xlarge", k));
+    WorkTaskDispatcher executor = new WorkTaskDispatcher((k,m)->Assertions.assertEquals("encoder.xlarge", k));
     executor.dispatch(task);
   }
   
