@@ -2,8 +2,8 @@ package com.creactiviti.piper.error;
 
 import java.util.Arrays;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import com.creactiviti.piper.core.error.ErrorHandler;
 import com.creactiviti.piper.core.error.ErrorHandlerChain;
@@ -17,8 +17,9 @@ public class ErrorHandlerChainTests {
   @Test
   public void test1() {
     ErrorHandler errorHandler = new ErrorHandler<Job>() {
+      @Override
       public void handle(Job j) {
-        Assert.assertEquals(SimpleJob.class, j.getClass()); 
+        Assertions.assertEquals(SimpleJob.class, j.getClass()); 
       }
     };
     ErrorHandlerChain chain = new ErrorHandlerChain(Arrays.asList(errorHandler));
@@ -28,13 +29,15 @@ public class ErrorHandlerChainTests {
   @Test
   public void test2() {
     ErrorHandler errorHandler1 = new ErrorHandler<Job>() {
+      @Override
       public void handle(Job j) {
         throw new IllegalStateException("should not get here");
       }
     };
     ErrorHandler errorHandler2 = new ErrorHandler<TaskExecution>() {
+      @Override
       public void handle(TaskExecution jt) {
-        Assert.assertEquals(SimpleTaskExecution.class, jt.getClass()); 
+        Assertions.assertEquals(SimpleTaskExecution.class, jt.getClass()); 
       }
     };
     ErrorHandlerChain chain = new ErrorHandlerChain(Arrays.asList(errorHandler1,errorHandler2));
