@@ -1,6 +1,6 @@
 package com.creactiviti.piper.error;
 
-import static org.mockito.Matchers.any;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.when;
 
 import java.util.Collections;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import com.creactiviti.piper.core.error.ErrorObject;
 import com.creactiviti.piper.core.error.TaskExecutionErrorHandler;
